first Fill attempt
While a key is pressed, the screen begins to fill with black. While no
key is pressed, it begins to fill with white.

I'm not sure if this is what they had in mind? As in: should I fill out
the whole screen as soon as a keypress is detected?

The book suggests that my pattern is okay. Will probably watch the
video to confirm.